dedupe basic API types from `modality-mutator-protocol`
