+++ client-libraries/rust/modality-api/src/lib.rs
@@ -3,3 +3,9 @@ pub mod protocol;
 pub mod types;
 
 pub use types::*;
+
+#[cfg(feature = "serde")]
+mod serde;
+
+#[cfg(any(test, feature = "proptest_strategies"))]
+pub mod proptest_strategies;
